Rename all reducer functions to reflect the state that they manage (convention)
