+++ client/reducers/currentItem.js
