+++ client/reducers/filteredItems.js
