+++ client/reducers/listedItems.js
